+++ NixieClockDori/DisplayTask_Flash.h
@@ -6,7 +6,7 @@ class DisplayTask_Flash : public DisplayTask {
 public:
 	DisplayTask_Flash();
 
-	bool refresh(uint16_t displayData[]) override;
+	bool refresh(byte displayData[]) override;
 
 	uint16_t symbols[6];
 	unsigned long endTime = 0;
