+++ NixieClockDori/DisplayTask_Stopwatch.h
@@ -6,7 +6,7 @@ class DisplayTask_Stopwatch : public DisplayTask {
 public:
 	DisplayTask_Stopwatch();
 
-	bool refresh(uint16_t displayData[]) override;
+	bool refresh(byte displayData[]) override;
 
 	void pause();
 	void resume();
